(#1587) Updates nci-header storybook with twig
Closes #1587
