+++ testing/ncids-css-testing/stories/components/nci-header/nci-header.default.stories.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import Component from '@nciocpl/ncids-twig/components/nci-header/nci-header.twig';
-import css from './nci-header.scss';
+import css from './index.scss';
 
 import { TestCase } from '../../../components/test-case';
 import { NCIExtendedHeaderWithMegaMenu } from '@nciocpl/ncids-js/nci-header';
 import { MockMegaMenuAdapter } from './mock-adapters/MockMegaMenuAdapter';
 import { MockMobileMenuAdapter } from './mock-adapters/MockMobileMenuAdapter';
-import { ExtendedContent, ExtendedSpanishContent } from './content';
+import { DefaultContent, DefaultSpanishContent } from './content';
 
 export default {
-	title: 'Components/Header/Variants/With Mega Menu',
+	title: 'Components/Header/Default',
 	args: {
 		ncidsInitJs: () => {
 			const header = document.querySelector('.nci-header');
@@ -24,9 +24,9 @@ export default {
 const Template = (args) => Component(args);
 
 export const English = () => (
-	<TestCase css={css} html={Template.bind({})(ExtendedContent)} />
+	<TestCase css={css} html={Template.bind({})(DefaultContent)} />
 );
 
 export const Spanish = () => (
-	<TestCase css={css} html={Template.bind({})(ExtendedSpanishContent)} />
+	<TestCase css={css} html={Template.bind({})(DefaultSpanishContent)} />
 );
