+++ testing/ncids-css-testing/stories/components/nci-header/nci-header.test-cases.default.stories.jsx
@@ -1,15 +1,19 @@
-import React from 'react';
 import Component from '@nciocpl/ncids-twig/components/nci-header/nci-header.twig';
-import css from './nci-header.scss';
+import css from './index.scss';
 
 import { TestCase } from '../../../components/test-case';
 import { NCIExtendedHeaderWithMegaMenu } from '@nciocpl/ncids-js/nci-header';
 import { MockMegaMenuAdapter } from './mock-adapters/MockMegaMenuAdapter';
 import { MockMobileMenuAdapter } from './mock-adapters/MockMobileMenuAdapter';
-import { DefaultContent, DefaultSpanishContent } from './content';
+import {
+	DefaultContent,
+	LongExtendedLinkContent,
+	ExtendedContentTwoItems,
+	ExtendedContentManyItems,
+} from './content';
 
 export default {
-	title: 'Components/Header/Default',
+	title: 'Components/Header/Test Cases/Default',
 	args: {
 		ncidsInitJs: () => {
 			const header = document.querySelector('.nci-header');
@@ -23,10 +27,21 @@ export default {
 
 const Template = (args) => Component(args);
 
-export const English = () => (
-	<TestCase css={css} html={Template.bind({})(DefaultContent)} />
+export const NoSearchField = () => (
+	<TestCase css={css} html={Template.bind({})({
+		...DefaultContent,
+		search_enabled: false,
+	})} />
 );
 
-export const Spanish = () => (
-	<TestCase css={css} html={Template.bind({})(DefaultSpanishContent)} />
+export const LongPrimaryLink = () => (
+	<TestCase css={css} html={Template.bind({})(LongExtendedLinkContent)} />
+);
+
+export const TwoItemNavigation = () => (
+	<TestCase css={css} html={Template.bind({})(ExtendedContentTwoItems)} />
+);
+
+export const TooManyItemsNavigation = () => (
+	<TestCase css={css} html={Template.bind({})(ExtendedContentManyItems)} />
 );
