+++ testing/ncids-css-testing/stories/components/nci-header/nci-header.variants.stories.jsx
@@ -1,19 +1,14 @@
-import React from 'react';
 import Component from '@nciocpl/ncids-twig/components/nci-header/nci-header.twig';
-import css from './nci-header.scss';
+import css from './index.scss';
 
 import { TestCase } from '../../../components/test-case';
 import { NCIExtendedHeaderWithMegaMenu } from '@nciocpl/ncids-js/nci-header';
 import { MockMegaMenuAdapter } from './mock-adapters/MockMegaMenuAdapter';
 import { MockMobileMenuAdapter } from './mock-adapters/MockMobileMenuAdapter';
-import {
-	DefaultContent,
-	ExtendedLongMegaMenuContent,
-	LongExtendedButtonContent,
-} from './content';
+import { ExtendedContent, ExtendedSpanishContent } from './content';
 
 export default {
-	title: 'Components/Header/Test Cases/Header with Mega Menu',
+	title: 'Components/Header/Variants/With Mega Menu',
 	args: {
 		ncidsInitJs: () => {
 			const header = document.querySelector('.nci-header');
@@ -23,24 +18,14 @@ export default {
 			});
 		},
 	},
-    nojsArgs:{
-
-    },
 };
 
 const Template = (args) => Component(args);
 
-const NoJSTemplate = (nojsArgs) => Component(nojsArgs);
-
-export const NoJS = () => (
-	<TestCase css={css} html={NoJSTemplate.bind({})(DefaultContent)} />
+export const English = () => (
+	<TestCase css={css} html={Template.bind({})(ExtendedContent)} />
 );
 
-export const LongPrimaryButton = () => (
-	<TestCase css={css} html={Template.bind({})(LongExtendedButtonContent)} />
+export const Spanish = () => (
+	<TestCase css={css} html={Template.bind({})(ExtendedSpanishContent)} />
 );
-
-export const LongMegaMenuContent = () => (
-	<TestCase css={css} html={Template.bind({})(ExtendedLongMegaMenuContent)} />
-);
-
